requirements.txt
torch>=2.4.0
triton>=3.0.0
transformers>=4.51.0
flash-attn
xxhash
aiohttp
fastapi
uvicorn
pandas
datasets
pillow
psutil
amd-quark==0.9.0
pynvml
seaborn
duckdb
pyzmq
zeus
